First two EventViewModelTests
