+++ AMEL/AMELTests/EventViewModelTests.swift
@@ -11,7 +11,7 @@ import CoreLocation
 import CoreData
 @testable import AMEL
 
-class AMELTests: XCTestCase {
+class EventFormatterTests: XCTestCase {
     
     class MockEvent: NSObject, EventProtocol {
         static func getEvents() -> NSFetchRequest<Event> {
@@ -32,7 +32,7 @@ class AMELTests: XCTestCase {
     
     let lm = LocationViewModel()
     let hlm = HeadingViewModel()
-    let formatter = LatLngFormatter(latitude: 0, longitude: 0)
+    
     override func setUp() {
         // Put setup code here. This method is called before the invocation of each test method in the class.
         
@@ -64,35 +64,7 @@ class AMELTests: XCTestCase {
         let eventResults = Event.getEvents()
         XCTAssertNotNil(eventResults)
     }
-    func testGetDMS() {
-        XCTAssertEqual("40-45.6", self.formatter.getDMS(number: 40.76))
-        XCTAssertEqual("73-59.0", self.formatter.getDMS(number: -73.984))
-    }
-    func testGetDMSRounding() {
-        XCTAssertEqual("73-58.5", self.formatter.getDMS(number: -73.974895))
-    }
-    func testLatZero() {
-        XCTAssertEqual("N00-00.0", self.formatter.formatLatitude(latitude: 0))
-    }
-    func testLatFormatN() {
-        XCTAssertEqual("N40-45.6", self.formatter.formatLatitude(latitude: 40.76))
-    }
-    func testLatFormatS() {
-        XCTAssertEqual("S21-02.9", self.formatter.formatLatitude(latitude: -21.047977))
-    }
-    func testLngZero() {
-        XCTAssertEqual("E00-00.0", self.formatter.formatLongitude(longitude: 0))
-    }
-    func testLngFormatW() {
-        XCTAssertEqual("W122-00.0", self.formatter.formatLongitude(longitude: -122))
-    }
-    func testLngFormatS() {
-        XCTAssertEqual("E122-00.0", self.formatter.formatLongitude(longitude: 122))
-    }
-    func testLatLngFormat() {
-        let sut = LatLngFormatter(latitude: 34.1622265, longitude: -119.0440182)
-        XCTAssertEqual("N34-09.7 W119-02.6", sut.getLatLng())
-    }
+    
     
     
     //Dont need performance testing at the moment
